update merge url and message
