+++ pages/profile/merge.js
@@ -3,25 +3,25 @@
 
 import { useEffect, useState } from 'react'
 import { useRouter } from 'next/router'
-import LoadingSpinner from '../components/LoadingSpinner'
-import ErrorDisplay from '../components/ErrorDisplay'
-import api from '../lib/api-client'
-import useLoginRedirect from '../hooks/useLoginRedirect'
-import SpinnerButton from '../components/SpinnerButton'
+import LoadingSpinner from '../../components/LoadingSpinner'
+import ErrorDisplay from '../../components/ErrorDisplay'
+import api from '../../lib/api-client'
+import useLoginRedirect from '../../hooks/useLoginRedirect'
+import SpinnerButton from '../../components/SpinnerButton'
 
-const Confirm = () => {
-  const { user, accessToken } = useLoginRedirect()
+const Merge = () => {
+  const { user } = useLoginRedirect()
   const router = useRouter()
   const [error, setError] = useState(null)
   const [isLoading, setIsLoading] = useState(false)
-  const [emailToConfirm, setEmailToConfirm] = useState(null)
+  const [profileIdToMergeFrom, setProfileIdToMergeFrom] = useState(null)
 
-  const confirmEmail = () => {
+  const confirmProfileMerge = () => {
     setIsLoading(true)
     api
-      .put(`/activatelink/${router.query.token}`)
-      .then(({ confirmedEmail }) => {
-        promptMessage(`Thank you for confirming your email ${confirmedEmail ?? ''}`)
+      .put(`/mergelink/${router.query.token}`)
+      .then(() => {
+        promptMessage(`Thank you for confirming the profile merge.`)
         setIsLoading(false)
         router.replace('/')
       })
@@ -34,20 +34,20 @@ const Confirm = () => {
   const getActivatable = async () => {
     try {
       const {
-        activatable: { groupId, username, action },
+        activatable: { action, duplicateProfile, username },
       } = await api.get(`/activatable/${router.query.token}`)
-      if (action !== 'alternate') {
+      if (action !== 'merge') {
         setError({ statusCode: 403, message: 'Token is invalid.' })
         return
       }
       if (!user?.profile?.usernames?.includes(username)) {
         setError({
           statusCode: 403,
-          message: 'You are not authorized to activate this email.',
+          message: 'You are not authorized to perform this merge.',
         })
         return
       }
-      setEmailToConfirm(groupId)
+      setProfileIdToMergeFrom(duplicateProfile)
     } catch (apiError) {
       setError({ statusCode: apiError.status, message: apiError.message })
     }
@@ -57,40 +57,42 @@ const Confirm = () => {
     if (!router.isReady) return
 
     if (!router.query.token) {
-      setError({ statusCode: 404, message: 'Activation token not found' })
+      setError({ statusCode: 404, message: 'Token is invalid' })
     }
-
     getActivatable()
   }, [router.isReady, router.query])
 
   if (error) return <ErrorDisplay statusCode={error.statusCode} message={error.message} />
-  if (!emailToConfirm) return <LoadingSpinner />
+  if (!user || !profileIdToMergeFrom) return <LoadingSpinner />
 
   return (
     <>
       <header>
-        <h1>Account Linking</h1>
+        <h1>Profile Merge</h1>
       </header>
       <p className="mt-4">
-        Click Confirm Email button below to confirm adding <strong>{emailToConfirm}</strong> to
-        your account.
+        Click Confirm Profile Merge button below to confirm merging{' '}
+        <a href={`/profile?id=${profileIdToMergeFrom}`} target="_blank" rel="noreferrer">
+          <strong>{profileIdToMergeFrom}</strong>
+        </a>{' '}
+        to your profile.
       </p>
       <p className="mb-4">
         Please contact{' '}
         <a href="mailto:info@openreview.net" target="_blank" rel="noreferrer">
           info@openreview.net
         </a>{' '}
-        if you didn&apos;t add this email.
+        if you didn&apos;t request for this profile merge.
       </p>
 
       <div className="response-buttons">
-        <SpinnerButton loading={isLoading} disabled={isLoading} onClick={confirmEmail}>
-          Confirm Email
+        <SpinnerButton loading={isLoading} disabled={isLoading} onClick={confirmProfileMerge}>
+          Confirm Profile Merge
         </SpinnerButton>
       </div>
     </>
   )
 }
 
-Confirm.bodyClass = 'confirm'
-export default Confirm
+Merge.bodyClass = 'confirm'
+export default Merge
